+++ src/blogmore/generator.py
@@ -256,6 +256,7 @@ class SiteGenerator:
             "with_read_time": self.site_config.with_read_time,
             "with_advert": self.site_config.with_advert,
             "default_author": self.site_config.default_author,
+            "extra_head_tags": self.site_config.head,
             "fontawesome_css_url": self._with_cache_bust(self._fontawesome_css_url),
             "fontawesome_woff2_url": FONTAWESOME_CDN_BRANDS_WOFF2_URL,
             "styles_css_url": styles_css_url,